Move ksrpc-core @KsrpcInternal symbols into `internal` package
Relocates the remaining cross-module internal symbols in ksrpc-core so
that their package already signals intent (`com.monkopedia.ksrpc.internal`),
matching the convention used by ksrpc-packets, ksrpc-jsonrpc, and
ksrpc-sockets. The `@KsrpcInternal` gating remains in place; this is a
package move only — implementations are unchanged.

- `ServiceExecutor` interface (was `com.monkopedia.ksrpc`)
- `ERROR_PREFIX` / `ENDPOINT_NOT_FOUND_PREFIX` constants (was
  `com.monkopedia.ksrpc`)
- `randomUuid()` expect/actual (was `com.monkopedia.ksrpc.channels`,
  actuals lived in `channels/UtilsJvm.kt`, `UtilsNative.kt`,
  `UtilsJs.kt`, and the wasmJs `UtilsJs.kt`)

Compiler plugin updated: `FqConstants.SERVICE_EXECUTOR` now points at
`com.monkopedia.ksrpc.internal.ServiceExecutor`, so the IR emitted for
generated service stubs references the new FQN. BCV api dumps updated
accordingly for ksrpc-core, ksrpc-flow, and ksrpc-introspection.
