+++ ksrpc-core/src/commonMain/kotlin/channels/SerializedChannel.kt
@@ -17,8 +17,6 @@
 
 package com.monkopedia.ksrpc.channels
 
-import com.monkopedia.ksrpc.ENDPOINT_NOT_FOUND_PREFIX
-import com.monkopedia.ksrpc.ERROR_PREFIX
 import com.monkopedia.ksrpc.KsrpcEnvironment
 import com.monkopedia.ksrpc.RpcMethod
 import com.monkopedia.ksrpc.RpcObject
@@ -27,6 +25,8 @@ import com.monkopedia.ksrpc.SuspendCloseableObservable
 import com.monkopedia.ksrpc.annotation.KsMethod
 import com.monkopedia.ksrpc.annotation.KsrpcInternal
 import com.monkopedia.ksrpc.channels.ChannelClient.Companion.DEFAULT
+import com.monkopedia.ksrpc.internal.ENDPOINT_NOT_FOUND_PREFIX
+import com.monkopedia.ksrpc.internal.ERROR_PREFIX
 import com.monkopedia.ksrpc.internal.HostSerializedServiceImpl
 import com.monkopedia.ksrpc.rpcObject
 import kotlin.coroutines.CoroutineContext
@@ -186,9 +186,6 @@ interface SerializedService<T> :
         callId: RpcCallId?
     ): CallData<T> = call(endpoint.endpoint, input, callId)
 }
-
-@KsrpcInternal
-expect fun randomUuid(): String
 
 /**
  * Wrapper around data being serialized through calls.
