+++ ksrpc-core/src/commonMain/kotlin/internal/ErrorPrefixes.kt
@@ -13,10 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.monkopedia.ksrpc.channels
+package com.monkopedia.ksrpc.internal
 
-import com.aventrix.jnanoid.jnanoid.NanoIdUtils
 import com.monkopedia.ksrpc.annotation.KsrpcInternal
 
 @KsrpcInternal
-actual fun randomUuid(): String = NanoIdUtils.randomNanoId()
+const val ERROR_PREFIX = "ERROR:"
+
+@KsrpcInternal
+const val ENDPOINT_NOT_FOUND_PREFIX = "ENDPOINT_NOT_FOUND:"
