+++ ksrpc-core/src/commonMain/kotlin/internal/RandomUuid.kt
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.monkopedia.ksrpc.channels
+package com.monkopedia.ksrpc.internal
 
 import com.monkopedia.ksrpc.annotation.KsrpcInternal
-import nanoid.NanoIdUtils
 
 @KsrpcInternal
-actual fun randomUuid(): String = NanoIdUtils.randomNanoId()
+expect fun randomUuid(): String
